pyproject.toml
[project]
name = "disserta-bench"
version = "0.1.0"
description = "Dataset of open-ended questions from Brazilian university entrance exams (FUVEST, ITA, Comvest), with official solutions and evaluation rubrics."
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "dspy>=3.1.3",
    "google-genai>=1.0.0",
    "ollama>=0.6.1",
    "pymupdf>=1.25.0",
]

[project.scripts]
download-fuvest = "scripts.download_fuvest:main"
extraction-pipeline = "scripts.extraction_pipeline:main"
run-pipeline = "scripts.run_pipeline:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["T201"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scripts"]

[tool.ty.environment]
python = ".venv"

[tool.ty.src]
include = ["./scripts"]
